add birth date to Case

+++ event/event-springdata/src/main/java/net/bakaar/sandbox/event/springdata/EventRaisedRepository.java
@@ -2,5 +2,5 @@ package net.bakaar.sandbox.event.springdata;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface EventRaisedRepository extends CrudRepository<DBEventRaised, Integer> {
+public interface EventRaisedRepository extends CrudRepository<DBEventRaised, Long> {
 }
